Reduce number of allocations per point
tcp (non Pickle) and udp receiver create single datapoint per line,
which is very likely not to be used as is, but be merged into
already existing one in cache.

This patch introduces new SinglePoint data structure which
requires a single heap allocation only - metric name string.

It is then passed on stack to cache, where either point values are
merged to existing points or new points structure is created only
if necessary.